% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersje_zmiennych_opisujacych_okresy.R
\name{data2rokszk}
\alias{data2rokszk}
\title{Zamiana daty w formacie "\%d-\%m-\%Y" na rok szkolny}
\usage{
data2rokszk(x, max, tekst = TRUE)
}
\arguments{
\item{x}{wektor ciągów znaków zawierający daty w formacie}

\item{max}{liczba - wartość odpowiadająca \strong{pierwszemu} rokowi roku
szkolnego, który zostanie przypisana datom opisującym niezakończony okres
nauki (tj. "31-12-9999")}

\item{tekst}{wartość logiczna - czy zwrócony powinien zostać tekst z pełnym
opisem roku szkolnego (np. "2023/2024"), czy liczba całkowita wskazującą
\strong{pierwszy} rok kalendarzowy w ramach odpowiedniego roku szkolnego}
}
\value{
wektor ciągów znaków lub wektor liczb całkowitych
}
\description{
Funkcja pomocnicza pozwalająca zamienić datę w formacie
"\%d-\%m-\%Y" na rok szkolny (jako liczbę wskazującą \strong{pierwszy} rok
kalendarzowy w ramach odpowiedniego roku szkolnego albo tekst z pełnym opisem
roku szkolnego).
}
\examples{
data2rokszk(c("30-08-2023", "01-09-2023", "31-12-9999"),
            max = 2024)
data2rokszk(c("30-08-2023", "01-09-2023", "31-12-9999"),
            max = 2024, tekst = FALSE)
}
\seealso{
\code{\link[=okres2rokmiesiac]{okres2rokmiesiac()}}, \code{\link[=okres2rokszk]{okres2rokszk()}}, \code{\link[=okres2data]{okres2data()}},
\code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}}
}
