% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_w19.R
\name{imputuj_wynagrodzenia}
\alias{imputuj_wynagrodzenia}
\title{Funkcje pomocnicze - przygotowanie danych do tabeli wejsciowej W19}
\usage{
imputuj_wynagrodzenia(w, zmiany)
}
\arguments{
\item{w}{wektor liczb - z założenia są to wartości przeciętnych miesięcznych
wynagrodzeń w powiecie w kolejnych okresach (latach); może zawierać braki
danych}

\item{zmiany}{wektor liczb - z założenia są to ilorazy przeciętnych
miesięcznych wynagrodzeń w danym okresie (roku) do przeciętnych miesięcznych
wynagrodzeń w poprzednim okresie (roku) w województwie, w którym znajduje się
dany powiat}
}
\value{
wektor liczb: \code{w} z dokonanymi imputacjami
}
\description{
Nieeksportowana funkcja wykorzystywana wewnętrznie przez
\code{\link[=przygotuj_dane_do_w19]{przygotuj_dane_do_w19()}} - przeprowadza imputację wartości wskaźnika na
podstawie informacji o zmianach przeciętnego poziomu wynagrodzeń.
}
\details{
Imputacja przeprowadzana jest krokowo i \strong{wymaga, aby pierwsza wartość \code{w}
nie była brakiem danych}. Przebiega w ten sposób, że dla każdego elementu
\code{w}, począwszy od drugiego, jeśli jest on brakiem danych, to podstawia się
za niego wartość iloczynu poprzedniego elementu \code{w} i danego elementu
\code{zmiany} (tj. arbitralnie zakłada się, że przeciętne zarobki w danym powiecie
zmieniły się procentowo względem poprzedniego roku o tyle samo, co w całym
województwie).
}
