% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersje_zmiennych_opisujacych_okresy.R
\name{okres2data}
\alias{okres2data}
\title{Zamiana "okresu" na date w formacie "\%d-\%m-\%Y"}
\usage{
okres2data(x, dzien = "01", warnIfNAs = TRUE)
}
\arguments{
\item{x}{liczba całkowita (co do zasady wartość kolumny \code{okres})}

\item{dzien}{opcjonalnie liczba całkowita lub ciąg znaków dający się
skonwertować na liczbę całkowitą opisująca dzień miesiąca, który zostanie
przypisany tworzonym datom; może być to też wektor takich liczb lub ciągów
znaków, o liczbie elementów równej liczbie elementów argumentu \code{x}, co
pozwala przypisać różne dni różnym okresom w ramach \code{x}}

\item{warnIfNAs}{wartość logiczna - czy generować ostrzeżenie, jeśli argument
\code{x} zawiera braki danych?}
}
\value{
wektor ciągów znaków z datami zapisanymi w formacie "\%d-\%m-\%Y"
}
\description{
Funkcja pomocnicza pozwalająca zamienić numer "okresu", używany
jako identyfikator miesiąca w tabelach \emph{pośrednich} na odpowiadającą mu
ciąg znaków opisujący datę w formacie "\%d-\%m-\%Y".
}
\examples{
okres2data(24249)
okres2data(c(24249, 24252, 24253))
}
\seealso{
\code{\link[=okres2rokmiesiac]{okres2rokmiesiac()}}, \code{\link[=okres2rokszk]{okres2rokszk()}}, \code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}},
\code{\link[=data2rokszk]{data2rokszk()}}
}
