% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersje_zmiennych_opisujacych_okresy.R
\name{okres2rokmiesiac}
\alias{okres2rokmiesiac}
\title{Zamiana "okresu" na rok i miesiac}
\usage{
okres2rokmiesiac(x, SIMPLIFY = TRUE, warnIfNAs = TRUE)
}
\arguments{
\item{x}{liczba całkowita (co do zasady wartość kolumny \code{okres})}

\item{SIMPLIFY}{wartość logiczna - jeśli argument \code{x} jest pojedynczą liczbą
(wektorem jednoelementowym), to czy zwrócony ma być dwuelementowy wektor
liczb zamiast listy?}

\item{warnIfNAs}{wartość logiczna - czy generować ostrzeżenie, jeśli argument
\code{x} zawiera braki danych?}
}
\value{
lista złożona z dwóch elementów: \code{rok} i \code{miesiac}, z których każdy
jest wektorem liczb całkowitych lub dwuelementowy wektor liczb całkowitych
(jeśli argument \code{SIMPLIFY= TRUE}, a argument \code{x} jest długości 1)
}
\description{
Funkcja pomocnicza pozwalająca zamienić numer "okresu", używany
jako identyfikator miesiąca w tabelach \emph{pośrednich} na odpowiadający mu
rok i numer miesiąca (w ramach roku).
}
\examples{
okres2rokmiesiac(24249)
okres2rokmiesiac(c(24249, 24252, 24253))
}
\seealso{
\code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}}, \code{\link[=okres2rokszk]{okres2rokszk()}}, \code{\link[=okres2data]{okres2data()}},
\code{\link[=data2rokszk]{data2rokszk()}}
}
