% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersje_zmiennych_opisujacych_okresy.R
\name{okres2rokszk}
\alias{okres2rokszk}
\title{Zamiana "okresu" na rok szkolny}
\usage{
okres2rokszk(x, tekst = TRUE, warnIfNAs = TRUE)
}
\arguments{
\item{x}{liczba całkowita (co do zasady wartość kolumny \code{okres})}

\item{tekst}{wartość logiczna - czy zwrócony powinien zostać tekst z pełnym
opisem roku szkolnego (np. "2023/2024"), czy liczba całkowita wskazującą
\strong{pierwszy} rok kalendarzowy w ramach odpowiedniego roku szkolnego}

\item{warnIfNAs}{wartość logiczna - czy generować ostrzeżenie, jeśli argument
\code{x} zawiera braki danych?}
}
\value{
wektor ciągów znaków lub wektor liczb całkowitych
}
\description{
Funkcja pomocnicza pozwalająca zamienić numer "okresu", używany
jako identyfikator miesiąca w tabelach \emph{pośrednich} na odpowiadający mu
rok szkolny (jako liczbę wskazującą \strong{pierwszy} rok kalendarzowy w ramach
odpowiedniego roku szkolnego albo tekst z pełnym opisem roku szkolnego).
}
\examples{
okres2rokszk(c(24248, 24249))
okres2rokszk(c(24248, 24249), tekst = FALSE)
}
\seealso{
\code{\link[=data2rokszk]{data2rokszk()}}, \code{\link[=okres2rokmiesiac]{okres2rokmiesiac()}}, \code{\link[=okres2data]{okres2data()}},
\code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}}
}
