% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/porownaj_liczbe_wierszy_w_tabelach_wejsciowych.R
\name{porownaj_liczbe_wierszy_w_tabelach_wejsciowych}
\alias{porownaj_liczbe_wierszy_w_tabelach_wejsciowych}
\title{Heurystyczna kontrola kompletnosci danych wejsciowych}
\usage{
porownaj_liczbe_wierszy_w_tabelach_wejsciowych(...)
}
\arguments{
\item{...}{ciągi znaków ze ścieżkami do innych, porównywanych katalogów}
}
\value{
Ramka danych z wierszami reprezentującymi poszczególne tabele
\emph{wejściowe} i kolumnami reprezentującymi poszczególne katalogi.
}
\description{
Funkcja przygotowuje zestawienie liczby wierszy w poszczególnych plikach
z tabelami \emph{wejściowymi} w podanych katalogach.
}
\examples{
\dontrun{
  w <- porownaj_liczbe_wierszy_w_tabelach_wejsciowych("./",
                                                      "../"dane 2024)
  print(w, n = Inf)
}
}
