% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_w19.R
\name{przygotuj_dane_do_w19}
\alias{przygotuj_dane_do_w19}
\title{Przygotowanie danych do tabeli wejsciowej W19}
\usage{
przygotuj_dane_do_w19(
  bezrobocie,
  wynagrodzenia,
  wynagrodzeniaWoj = NULL,
  zapiszDoPliku = "W19.csv"
)
}
\arguments{
\item{bezrobocie}{ramka danych zwrócona przez wywołanie
\code{MLASZdane::pobierz_dane_bdl(MLASZdane::znajdz_wskazniki_bdl("stopa bezrobocia rejestrowanego"), zakresLat, "powiaty")}}

\item{wynagrodzenia}{ramka danych zwrócona przez wywołanie
\code{MLASZdane::pobierz_dane_bdl(MLASZdane::znajdz_wskazniki_bdl("przeciętne miesięczne wynagrodzenia brutto"), zakresLat, "powiaty")}}

\item{wynagrodzeniaWoj}{opcjonalnie ramka danych zwrócona przez wywołanie
\code{MLASZdane::pobierz_dane_bdl(MLASZdane::znajdz_wskazniki_bdl("przeciętne miesięczne wynagrodzenia brutto"), zakresLat, "województwa")}}

\item{zapiszDoPliku}{opcjonalnie ciąg znaków z nazwą pliku CSV, do którego ma
zostać zapisane przygotowane zestawienie; aby uniknąć zapisu do pliku, jako
wartość argumentu należy podać \code{NULL}}
}
\value{
ramka danych w formacie odpowiadającym specyfikacji
tabeli \emph{wejściowej} W19; jeśli nastąpił zapis do pliku CSV, wynik zwracany
jest jako \emph{ukryty}
}
\description{
Funkcja przekształca zestawienia wskaźników bezrobocia
i przeciętnych (w skali roku) miesięcznych wynagrodzeń, przekazane do niej
w formie, w jakiej są zwracane przez \link[MLASZdane:pobierz_dane_bdl]{MLASZdane::pobierz_dane_bdl} do formatu
wymaganego dla tabeli \emph{wejściowej} W19. O ile podano argument
\code{wynagrodzeniaWoj}, próbuje też zaimputować ew. braki danych we wskaźniku
przeciętnych wynagrodzeń występujące na poziomie konkretnych powiatów.
}
\examples{
\dontrun{
library(MLASZdane)
bezrobocie <- znajdz_wskazniki_bdl("stopa bezrobocia rejestrowanego") \%>\%
  pobierz_dane_bdl(2019:2024, "powiaty")
wynagrodzenia <- wskaznik_bdl(64428, "przeciętne miesięczne wynagrodzenia brutto") \%>\%
  pobierz_dane_bdl(2019:2024, "powiaty")
wynagrodzeniaWoj <- wskaznik_bdl(64428, "przeciętne miesięczne wynagrodzenia brutto") \%>\%
  pobierz_dane_bdl(2019:2024, "województwa")

przygotuj_dane_do_w19(bezrobocie, wynagrodzenia, wynagrodzeniaWoj,
                      zapiszDoPliku = "W19.csv")
}
}
\seealso{
\code{\link[MLASZdane:pobierz_dane_bdl]{MLASZdane::pobierz_dane_bdl()}}, \code{\link[MLASZdane:znajdz_wskazniki_bdl]{MLASZdane::znajdz_wskazniki_bdl()}}
}
