% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_w26.R
\name{przygotuj_dane_do_w26}
\alias{przygotuj_dane_do_w26}
\title{Przygotowanie danych do tabeli wejsciowej W26}
\usage{
przygotuj_dane_do_w26(
  rokMonitoringu,
  rspo = list.files(pattern = "^rspo_.*\\\\.csv$"),
  idOrgProw = "id-organow-prowadzacych.csv",
  zapiszDoPliku = "W26.csv"
)
}
\arguments{
\item{rokMonitoringu}{rok (edycja) prowadzenie monitoringu (liczba)}

\item{rspo}{opcjonalnie wektor ciągów znaków z nazwami plików CSV
zawierającymi dane wyeksportowane z (publicznego API WWW) RSPO (lub pozyskane
w inny sposób, ale zapisane w tej samej strukturze)}

\item{idOrgProw}{opcjonalnie nazwa pliku zawierającego dane o \emph{idPodmiotu}
SIO organów prowadzących szkół, wyeksportowane z hurtowni danych SIO; można
też podać \code{NULL}, aby nie dołączać tych danych (w szczególności, gdy nie są
dostępne)}

\item{zapiszDoPliku}{opcjonalnie ciąg znaków z nazwą pliku CSV, do którego ma
zostać zapisane przygotowane zestawienie; aby uniknąć zapisu do pliku, jako
wartość argumentu należy podać \code{NULL}}
}
\value{
ramka danych w formacie odpowiadającym specyfikacji
tabeli \emph{wejściowej} W26; jeśli nastąpił zapis do pliku CSV, wynik zwracany
jest jako \emph{ukryty}
}
\description{
Funkcja przekształca dane o szkołach wyeksportowane
z (publicznego API WWW) RSPO i, opcjonalnie, dane o \emph{idPodmiotu} SIO organów
prowadzących szkół, wyeksportowane z hurtowni danych SIO do struktury, jaka
jest oczekiwana w tabeli \emph{wejściowej} W26. Sprawdza przy tym, czy w danych
z RSPO znajdują się wszystkie szkoły, które występują w danych z danej edycji
monitoringu (plikach \emph{W2.csv}, \emph{W3.csv}, \emph{W4.csv} i \emph{W5.csv})
}
