% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_w19.R
\name{przygotuj_komunikat_o_brakach_powiaty}
\alias{przygotuj_komunikat_o_brakach_powiaty}
\title{Funkcje pomocnicze - przygotowanie danych do tabeli wejsciowej W19}
\usage{
przygotuj_komunikat_o_brakach_powiaty(brakiPowiaty, poImputacji = FALSE)
}
\arguments{
\item{brakiPowiaty}{ramka danych zwrócona przez wywołanie
\code{\link[=przygotuj_zestawienie_brakow_danych]{przygotuj_zestawienie_brakow_danych()}} z argumentem \code{co="powiaty"}}

\item{poImputacji}{wartość logiczna wskazująca, czy komunikat ma być
wyświetlony już po dokonaniu próby imputacji braków wartości wskaźników}
}
\value{
ciąg znaków z komunikatem
}
\description{
Nieeksportowana funkcja wykorzystywana wewnętrznie przez
\code{\link[=przygotuj_dane_do_w19]{przygotuj_dane_do_w19()}} - przygotowuje komunikat o brakach wartości
wskaźników bezrobocia i przeciętnych wynagrodzeń, które dotyczą
poszczególnych powiatów.
}
