% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_tabele_posrednie.R
\name{przygotuj_tabele_posrednie}
\alias{przygotuj_tabele_posrednie}
\title{Przygotowywanie tabel "posrednich"}
\usage{
przygotuj_tabele_posrednie(
  baza,
  rokMonitoringu,
  okresOd = 0L,
  okresDo = rokmiesiac2okres(rokMonitoringu, 3L),
  okresyP2 = seq(rokmiesiac2okres(rokMonitoringu - 5L, 12L),
    rokmiesiac2okres(rokMonitoringu - 1L, 12L), by = 6L),
  imputujDateKoncaSzkoly = as.Date(paste0(rokMonitoringu, "-08-31")),
  imputujDateKoncaStudiow = as.Date(paste0(rokMonitoringu, "-09-30")),
  imputujDateKoncaAdresu = as.Date(paste0(rokMonitoringu, "-05-31")),
  minDniEdukacjiWMiesiacu = 14L,
  przygotujP1 = TRUE,
  przygotujP2 = TRUE,
  przygotujP3 = TRUE,
  przygotujP4 = przygotujP3,
  przygotujP5 = TRUE,
  przygotujP6 = TRUE
)
}
\arguments{
\item{baza}{uchwyt połączenia do bazy lub lista argumentów do funkcji
\code{\link[DBI:dbConnect]{DBI::dbConnect()}} umożliwiających nawiązanie połączenia z bazą
danych, w której mają zostać zapisane wczytywane dane}

\item{rokMonitoringu}{rok, w którym prowadzony jest monitoring (ma znaczenie
dla ustalenia adekwatnych przedziałów czasu, dla których mają zostać
obliczone wskaźniki, a na podstawie samych danych funkcja nie ma jak
go odgadnąć, nie narażając się na błędy)}

\item{okresOd}{opcjonalnie liczba całkowita - liczbę miesięcy, o którą
początek okresu objętego wskaźnikami w tabeli P3 ma być
\strong{przesunięty względem} stycznia roku ukończenia szkoły (zostania
absolwentem); domyślnie 0 (tzn. pierwszym miesiącem uwzględnionym w P3 jest
styczeń roku ukończenia szkoły)}

\item{okresDo}{opcjonalnie liczba całkowita - identyfikator miesiąca w formie
\code{12*rok+numer_miesiaca_w_ramach_roku} (p. \code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}}), który ma
być ostatnim miesiącem objętym wskaźnikami w tabeli P3; domyślnie marzec roku
prowadzenia monitoringu}

\item{okresyP2}{opcjonalnie wektor liczb całkowitych - identyfikatory
miesięcy w formie \code{12*rok+numer_miesiaca_w_ramach_roku}
(p. \code{\link[=rokmiesiac2okres]{rokmiesiac2okres()}}), dla których mają zostać obliczone wskaźniki
zwracane w tabeli P2 (tj. wskaźniki dotyczące form kontynuacji nauki);
domyślnie grudzień każdego roku objętego monitoringiem}

\item{imputujDateKoncaSzkoly}{opcjonalnie data, która ma zostać użyta jako
data zakończenia nauki w szkołach, w których absolwenci kontynuują edukację,
jeśli w danych (tabeli W3) jest ona nieokreślona (tzn. absolwent cały czas
się w danej szkole uczy); domyślnie koniec roku szkolnego kończącego się
w roku prowadzenia monitoringu}

\item{imputujDateKoncaStudiow}{opcjonalnie data, która ma zostać użyta jako
data zakończenia nauki na studiach, na których absolwenci kontynuują
edukację, jeśli w danych (tabeli W12) jest ona nieokreślona (tzn. absolwent
cały czas tam studiuje); domyślnie koniec roku akademickiego kończącego się
w roku prowadzenia monitoringu}

\item{imputujDateKoncaAdresu}{opcjonalnie data, która ma zostać użyta jako
data końca obowiązywania danego adresu do korespondencji, zamieszkania lub
zameldowania (z danych ZUS), jeśli w danych (tabeli W15) jest ona
nieokreślona; domyślnie ostatni dzień maja roku prowadzenia monitoringu, tj.
koniec okresu objętego eksportem danych z ZUS}

\item{minDniEdukacjiWMiesiacu}{opcjonalnie liczba całkowita - minimalna
liczba dni nauki w miesiącu, wystarczająca aby wskaźniki opisujące status
kontynuacji edukacji w danym miesiącu w tabeli P3 wskazywały na kontynuację
nauki (a nie na jej brak); domyślnie 14}

\item{przygotujP1}{opcjonalnie wartość logiczna - czy przygotować
tabelę \emph{pośrednią} P1? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji)}

\item{przygotujP2}{opcjonalnie wartość logiczna - czy przygotować
tabelę \emph{pośrednią} P2? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji)}

\item{przygotujP3}{opcjonalnie wartość logiczna - czy przygotować
tabelę \emph{pośrednią} P3? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji)}

\item{przygotujP4}{opcjonalnie wartość logiczna - czy przygotować
\emph{tabelę pośrednią} P4? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji); aby tabela P4 mogła zostać
przygotowana, konieczne jest również przygotowanie tabeli P3}

\item{przygotujP5}{opcjonalnie wartość logiczna - czy przygotować
tabelę \emph{pośrednią} P5? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji); aby tabela P5 mogła zostać
przygotowana, konieczne jest również przygotowanie tabeli P3}

\item{przygotujP6}{opcjonalnie wartość logiczna - czy przygotować
tabelę \emph{pośrednią} P6? (umożliwia wyłączenie przygotowania wybranej
tabeli w ramach danego wywołania funkcji);}
}
\value{
lista ramek danych z przygotowanymi tabelami
}
\description{
Na podstawie zawartych w bazie danych z tabel wejściowych
funkcja przygotowuje 6 tabel \emph{pośrednich}, które zawierają zestawienia
wskaźników (bardziej) bezpośrednio użytecznych analitycznie. Aby zapisać tak
przygotowane tabele do relacyjnej bazy danych, należy następnie użyć funkcji
\code{\link[=wczytaj_tabele_posrednie]{wczytaj_tabele_posrednie()}}.
}
\details{
Dane zawarte w bazie powinny obejmować \strong{tylko jeden rok
prowadzenia monitoringu}, ale mogą obejmować \strong{kilka różnych okresów od
ukończenia szkoły}.

Wskaźniki dotyczące kontynuacji nauki w tabeli P3 nie odnoszą się do nauki
w szkole, jako absolwent której dana osoba została objęta monitoringiem,
a jedynie do \emph{kontynuacji} nauki (w praktyce: nauki w innej szkole niż
ta, jako absolwent której dana osoba została objęta monitoringiem).

W przypadku KKZ i KUZ przy kodowaniu wskaźników opisujących status
kontynuacji nauki w tabeli P3 lub branżę kontynuacji nauki w tabeli P2
zostały wzięte pod uwagę tylko kursy, które mają podaną datę zakończenia
(co do tych, które nie mają jej podanej, nie da się rzetelnie rozstrzygać,
czy zostały one ukończone, czy zarzucone).
}
\seealso{
\code{\link[=wczytaj_tabele_posrednie]{wczytaj_tabele_posrednie()}}
}
