% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_do_w19.R
\name{przygotuj_zestawienie_brakow_danych}
\alias{przygotuj_zestawienie_brakow_danych}
\title{Funkcje pomocnicze - przygotowanie danych do tabeli wejsciowej W19}
\usage{
przygotuj_zestawienie_brakow_danych(wskazniki, co = c("wskazniki", "powiaty"))
}
\arguments{
\item{wskazniki}{ramka danych w formacie odpowiadającym specyfikacji
\emph{tabeli wejściowej} W19}

\item{co}{ciąg znaków wskazujący, czy ma zostać przygotowane zestawienie
braków danych opisujące, dla jakich lat i miesięcy brak wartości
poszczególnych wskaźników dla wszystkich powiatów w przekazanych danych
("wskazniki") lub dla jakich poszczególnych powiatów brak wartości
wskaźników, choć dla innych powiatów w tym samym roku i miesiącu są one
dostępne w przekazanych danych ("powiaty")}
}
\value{
ramka danych z zestawieniem brakujących wartości wskaźników
}
\description{
Nieeksportowana funkcja wykorzystywana wewnętrznie przez
\code{\link[=przygotuj_dane_do_w19]{przygotuj_dane_do_w19()}} - przygotowuje zestawienie braków wartości
wskaźników bezrobocia i przeciętnych wynagrodzeń.
}
