% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przytnij_wynagrodzenia.R
\name{przytnij_wynagrodzenia}
\alias{przytnij_wynagrodzenia}
\title{Przycinanie wartosci odstajacych oszacowan wynagrodzen}
\usage{
przytnij_wynagrodzenia(
  dane,
  zmKryterium,
  zmZastosuj = vector(mode = "character", length = 0),
  zmGrupujace = vector(mode = "character", length = 0),
  q = c(0.005, 0.995),
  usun0 = TRUE,
  zachowajOryginalne = TRUE
)
}
\arguments{
\item{dane}{ramka danych, typowo tabela \emph{pośrednia} P3}

\item{zmKryterium}{ciąch znaków - nazwa zmiennej, na podstawie rozkładu
której mają zostać obliczone progi}

\item{zmZastosuj}{opcjonalnie wektor ciągów znaków - nazwy zmiennych, których
wartości mają zostać przycięte (oprócz zmiennej podanej argumentem
\code{zmKryterium}, której wartości są przycinane zawsze)}

\item{zmGrupujace}{opcjonalnie wektor ciągów znaków - nazwy zmiennych,
przecięcie wartości których definiuje grupy, w ramach których mają zostać
obliczone progi}

\item{q}{opcjonalnie dwuelamentowy wektor liczbowy z wartościami kwantyli,
definiującymi progi odcięcia (domyślnie 0.005 i 0.995); jeśli zostanie podana
tylko jedna wartość, drugi kwantyl zostanie dobrany automatycznie, jako
jej dopełnienie do 1}

\item{usun0}{opcjonalnie wartość logiczna - czy przed obliczeniemn kwantyli
oraz dokonaniem przycięcia wartości przekształcanych zmiennych równe 0
powinny zostać zamienione na braki danych? (domyślnie \code{TRUE})}

\item{zachowajOryginalne}{opcjonalnie wartość logiczna - czy w danych
powinny zostać zachowane oryginalne, nieprzycięte kopie przekształcanych
zmiennych? (domyślnie \code{TRUE})
\strong{kopie takie otrzymają w nazwie przyrostek "_oryg"}}
}
\value{
ramka danych przekazana argumentem \code{dane} zawierająca przekształcone
zmienne i ew. te same zmienne nieprzekształcone, ale z dodanym do nazwy
przyrostkiem "o_ryg"
}
\description{
Funkcja przycina wartości odstające zmiennych - typowo
opisujących wynagrodzenia. Progi przycięcia obliczane są jako kwantyle zadane
argumentem \code{q} w odniesieniu do zmiennej o nazwie podaje argumentem
\code{zmKryterium}, oddzielnie w ramach każdej z podgrup wyróżnionych przez
przecięcie ze sobą wartości zmiennych, których nazwy zostały podane
argumentem \code{zmGrupujące}. Następnie wszystkie wartości zmiennych, których,
nazwy zostały podane argumentami \code{zmKryterium} lub \code{zmZastosuj}
przekształcane są tak, że wartości na zewnątrz od obliczonych progów
zamieniane są na wartości progów.
}
