% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersje_zmiennych_opisujacych_okresy.R
\name{rokmiesiac2okres}
\alias{rokmiesiac2okres}
\title{Obliczanie "okresu" na podstawie roku i miesiąca}
\usage{
rokmiesiac2okres(rok, miesiac, warnIfNAs = FALSE)
}
\arguments{
\item{rok}{wektor liczb całkowitych opisujących lata}

\item{miesiac}{wektor liczb całkowitych ze zbioru \{1, ..., 12\} opisujących
miesiące}

\item{warnIfNAs}{wartość logiczna - czy generować ostrzeżenie, jeśli argument
\code{x} zawiera braki danych?}
}
\value{
wektor liczb całkowitych
}
\description{
Funkcja pomocnicza pozwalająca obliczyć numer "okresu" na
na podstawie podanych lat i miesięcy.
}
\examples{
rokmiesiac2okres(c(2023, 2024), c(6, 12))
rokmiesiac2okres(c(2023, 2024), 12)
}
\seealso{
\code{\link[=okres2rokmiesiac]{okres2rokmiesiac()}}, \code{\link[=okres2rokszk]{okres2rokszk()}}, \code{\link[=okres2data]{okres2data()}},
\code{\link[=data2rokszk]{data2rokszk()}}
}
