% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele_wejsciowe.R
\name{tabele_wejsciowe}
\alias{tabele_wejsciowe}
\title{Struktura tabel wejsciowych}
\usage{
tabele_wejsciowe()
}
\value{
lista wektorów tekstowych
}
\description{
Funkcja zwraca listę zawierającą nazwy kolumn, które oczekiwane
są w poszczególnych plikach z tabelami wejściowymi.
}
