% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usun_duplikaty.R
\name{usun_duplikaty}
\alias{usun_duplikaty}
\title{Usuwanie zduplikowanych absolwentow z tabel "posrednich"}
\usage{
usun_duplikaty(x, p4 = NULL, duplikatWSzkole = TRUE, duplikatWieleSzkol = TRUE)
}
\arguments{
\item{x}{lista tabel pośrednich zwrócona przez \code{\link[=przygotuj_tabele_posrednie]{przygotuj_tabele_posrednie()}}
lub pojedyncza tabela \emph{pośrednia} (tj. ramka danych będąca pojedynczym
elementem takiej listy)}

\item{p4}{jeśli \code{x} jest pojedynczą tabelą \emph{pośrednią}, ramka danych
z tabelą \emph{pośrednią} P4 (na podstawie której zostaną zidentyfikowane
duplikaty); jeśli \code{x} jest listą, \code{NULL}}

\item{duplikatWSzkole}{wartość logiczna - czy usuwać absolwentów, którzy
występują kilkakrotnie w ramach tej samej szkoły?}

\item{duplikatWieleSzkol}{wartość logiczna - czy usuwać absolwentów, którzy
występują w kilku różnych szkołach?}
}
\value{
lista ramek danych lub ramka danych - w zależności od tego, jaki
obiekt został przekazany argumentem \code{x}
}
\description{
Funkcja usuwa zduplikowanych absolwentów - identyfikowanych na
podstawie tabeli \emph{pośredniej} P4 - z tabel \emph{pośrednich}.
}
\details{
\itemize{
\item \emph{Absolwent} identyfikowany jest przez kombinację wartości kolumn \code{id_abs}
i \code{rok_abs}.
\item Jeśli argumentem \code{x} przekazywana jest pojedyncza ramka danych z tabelą P4,
to należy tę samą ramkę danych przekazać również argumentem \code{p4}.
\item Tabela \emph{pośrednia} \strong{P6} - jeśli jest elementem przekazanej listy - zostanie
zwrócona \strong{niezmodyfikowana}.
}
}
