% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uzupelnij_wzgledne_wynagrodzenia.R
\name{uzupelnij_wzgledne_wynagrodzenia}
\alias{uzupelnij_wzgledne_wynagrodzenia}
\title{Uzupelnianie wartosci wskaznikow wzglednych wynagrodzen}
\usage{
uzupelnij_wzgledne_wynagrodzenia(
  p4,
  p3,
  plikW19 = "W19.csv",
  nieNadpisuj = TRUE
)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4, musi zawierać co najmniej
kolumny \code{id_abs} i \code{rok_abs});
jeśli w wyniku działania funkcji chce się uzyskać wyłącznie tabelę
\emph{pośrednią} P3 z uzupełnionymi wartościami kolumny \code{powiat_sr_wynagrodzenie}
oraz \code{powiat_bezrobocie}, można ustawić wartość tego argumentu na \code{NULL}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3, musi zawierać co najmniej
kolumny \code{id_abs}, \code{rok_abs}, \code{okres}, \code{wynagrodzenie_uop}, \code{teryt_zam}
i \code{nauka2}}

\item{plikW19}{opcjonalnie ciąg znaków z nazwą pliku CSV z tabelą \emph{wejściową}
W19 (p. \code{\link[=przygotuj_dane_do_w19]{przygotuj_dane_do_w19()}})}

\item{nieNadpisuj}{opcjonalnie wartość logiczna - czy uzupełnić wyłącznie te
wartości kolumny \code{powiat_sr_wynagrodzenie} w ramce danych przekazanej
argumentem \code{p3}, które są brakami danych (pozostawiając jak są te, które
mają już konkretne wartości)?}
}
\value{
ramka danych przekazana argumentem \code{p4}, w której na nowo
przeliczone zostały wartości kolumn \code{sr_wynagr_uop_nauka_r0_wrzgru}
i \code{sr_wynagr_uop_bez_nauki_r0_wrzgru} (lub, jeśli tych kolumn w niej nie
było, to zostały do niej dodane) albo ramka danych przekazana argumentem
\code{p3}, w której uzupełniono wartości kolumny \code{powiat_sr_wynagrodzenie} oraz
\code{powiat_bezrobocie} (jeśli funkcję wywołano z argumentem \code{p4} ustawionym na
\code{NULL})
}
\description{
Funkcja pozwala uzupełnić wartości wskaźników względnych
wynagrodzen (\code{sr_wynagr_uop_nauka_r0_wrzgru}
i \code{sr_wynagr_uop_bez_nauki_r0_wrzgru}) w tabeli \emph{pośredniej} P4, jeśli
wcześniej przygotowane tabele \emph{pośrednie} miały w nich braki danych dlatego,
że w momencie ich przygotowywania GUS nie zdążył jeszcze opublikować
wskaźnika średnich miesięcznych wynagrodzeń w powiatach dla ostatniego roku,
a teraz są już one dostępne (i została przygotowana zaktualizowana,
obejmująca je, wersja pliku \emph{wejściowego} W19.csv - p.
\code{\link[=przygotuj_dane_do_w19]{przygotuj_dane_do_w19()}})
}
\examples{
\dontrun{
library(MLASZdane)
# w katalogu roboczym musi znajdować się plik "W19.csv"!
p4 <- uzupelnij_wzgledne_wynagrodzenia(p4, p3)
p3 <- uzupelnij_wzgledne_wynagrodzenia(NULL, p3)
}
}
\seealso{
\code{\link[LOSYwskazniki:dodaj_wskazniki_prace]{LOSYwskazniki::dodaj_wskazniki_prace()}}
}
