% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_tabele_wejsciowe.R
\name{wczytaj_tabele_wejsciowe}
\alias{wczytaj_tabele_wejsciowe}
\title{Wczytywanie danych do bazy z plikow wejsciowych}
\usage{
wczytaj_tabele_wejsciowe(
  baza,
  folder = ".",
  wczytajDoBazy = TRUE,
  zapiszProblemy = TRUE,
  usunAbsWKilkuZaw = TRUE,
  plikLogu = NULL,
  nieWczytujDanychZUS = FALSE
)
}
\arguments{
\item{baza}{uchwyt połączenia do bazy lub lista argumentów do funkcji
\code{\link[DBI:dbConnect]{DBI::dbConnect()}} umożliwiających nawiązanie połączenia z bazą
danych, w której mają zostać zapisane wczytywane dane lub \code{NULL}, jeśli
argument \code{wczytajDoBazy} podano jako \code{FALSE}}

\item{folder}{opcjonalnie ścieżka do folderu zawierającego pliki CSV z danymi
(domyślnie używany jest aktywny katalog roboczy)}

\item{wczytajDoBazy}{opcjonalnie wartość logiczna - czy przeprowadzić zapis
do bazy? (jeśli \code{FALSE}, funkcja tylko wczyta pliki CSV i zaraportuje
wykryte w nich problemy)}

\item{zapiszProblemy}{opcjonalnie wartość logiczna - czy funkcja ma zapisywać
zestawienia wykrytych we wczytanych plikach problemów w formie plików CSV?}

\item{usunAbsWKilkuZaw}{czy absolwenci, którzy ukończyli kilka różnych
zawodów w tej samej szkole, \strong{niebędącej szkołą policealną} powinni
zostać wykluczeni z importu? (W edycjach monitoringu \strong{2021 i 2022}
tacy absolwenci \strong{nie} byli usuwani. Od edycji \strong{2023} są jednak
usuwani.)}

\item{plikLogu}{opcjonalnie nazwa pliku, do którego ma być zapisany log
wywołania funkcji}

\item{nieWczytujDanychZUS}{opcjonalnie wartość logiczna pozwalająca pominąć
wczytywanie do bazy danych z ZUS (i tym samym wydatnie skrócić proces) -
przydatna, jeśli proces ma na celu tylko wprowadzenie jakichś poprawek do
już wcześniej przygotowanych tabel \emph{pośrednich} w zakresie danych innych,
niż te z ZUS}
}
\value{
\code{NULL}
}
\description{
Funkcja szuka w podanym folderze plików CSV z danymi
z tabel \emph{wejściowych}, wczytuje je, kontroluje ich spójność i zapisuje
do bazy.
}
\details{
Dane wejściowe powinny obejmować \strong{tylko jeden rok prowadzenia
monitoringu}, ale mogą obejmować \strong{kilka różnych okresów od ukończenia
szkoły}.
}
\seealso{
\code{\link[=tabele_wejsciowe]{tabele_wejsciowe()}}
}
