% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zidentyfikuj_szkolo_lata_w_danych.R
\name{zidentyfikuj_szkolo_lata_w_danych}
\alias{zidentyfikuj_szkolo_lata_w_danych}
\title{Przygotowanie danych do tabeli wejsciowej W26}
\usage{
zidentyfikuj_szkolo_lata_w_danych(w2, w3, w4, w5, rokMonitoringu)
}
\arguments{
\item{w2}{ramka danych: tabela \emph{wejściowa} W2}

\item{w3}{ramka danych: tabela \emph{wejściowa} W3}

\item{w4}{ramka danych: tabela \emph{wejściowa} W4}

\item{w5}{ramka danych: tabela \emph{wejściowa} W5}

\item{rokMonitoringu}{rok (edycja) prowadzenie monitoringu (liczba)}
}
\value{
ramka danych zawierająca kolumny \code{ID_SZK} i \code{ROK_SZK}
}
\description{
Funkcja przygotowuje zestawienie unikalnych wartości id RSPO
szkoły i roku szkolnego występujące w danych z monitoringu (tj w tabelach
\emph{wejściowych} W2, W3, W4 i W5).
}
\seealso{
\code{\link[=wczytaj_tabele_wejsciowe]{wczytaj_tabele_wejsciowe()}}, \code{\link[=przygotuj_dane_do_w26]{przygotuj_dane_do_w26()}}
}
