% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okres2rokmiesiac.R
\name{okres2rokmiesiac}
\alias{okres2rokmiesiac}
\title{Zamiana "okresu" na rok i miesiac}
\usage{
okres2rokmiesiac(x)
}
\arguments{
\item{x}{liczba całkowita (co do zasady wartość kolumny `okres`)}
}
\value{
dwuelementowy wektor liczb całkowitych, z których pierwsza opisuje
rok, a druga numer miesiąca (w ramach roku)
}
\description{
Funkcja pomocnicza pozwalająca zamienić numer "okresu", używany
jako identyfikator miesiąca w tabelach \emph{pośrednich} na odpowiadający mu
rok i numer miesiąca (w ramach roku).
}
\examples{
okres2rokmiesiac(24249)
}
