% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_tabele_posrednie.R
\name{wczytaj_tabele_posrednie}
\alias{wczytaj_tabele_posrednie}
\title{Wczytanie tabel "posrednich" do relacyjnej bazy danych}
\usage{
wczytaj_tabele_posrednie(tabelePosrednie, baza)
}
\arguments{
\item{tabelePosrednie}{lista ramek danych zwrócona przez funkcję
\code{\link{przygotuj_tabele_posrednie}} (może zawierać wszystkie 4 tabele
\emph{p1}, \emph{p2}, \emph{p3} i \emph{p4}, lub tylko niektóre z nich)}

\item{baza}{uchwyt połączenia do bazy lub lista argumentów do funkcji
\code{\link[DBI]{dbConnect}} umożliwiających nawiązanie połączenia z bazą
danych, w której mają zostać zapisane wczytwane dane}
}
\description{
Funkcja wczytuje do relacyjnej bazy danych zawartość tabel
\emph{pośrednich}, które zostały wcześniej przygotowane jako efekt działania
funkcji \code{\link{przygotuj_tabele_posrednie}}.
}
